requirements.txt
numpy
pandas
tqdm
scipy
scikit-learn
biopython
einops
torch
transformers
